Refactor GallifreyDB to delegate temporal queries to HistoricalStorage
- Extract temporal query logic (get_at_time, history, diff) from `src/db.rs` to `src/storage/historical/mod.rs`.
- Move shared result types (`VersionInfo`, `EntityHistory`, `VersionDiff`) to `src/core/history.rs` to avoid circular dependencies.
- Update `HistoricalStorage` accessors to verify visibility of indexed versions for robustness.
- Reduce `src/db.rs` size and complexity by delegating to `self.historical`.
- Re-export moved types in `src/query/result.rs` to preserve public API.
